Move createParquetMetadata to ParquetMetadata
